+++ packages/ub-pub/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+- fixed missed localization for `Select device` caption on `@ub-d/iit-sign-web/IitReadPk.vue` form 
 
 ## [5.5.16] - 2020-07-19
 ### Added