+++ packages/ub-pub/CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
  - fixed missed localization for `Select device` caption on `@ub-d/iit-sign-web/IitReadPk.vue` form
  - prevent sets of `connection.domain` to null in case `onGotApplicationDomain` handler is passed to `AsyncConection.connect` and
-   implementation do not return domain  
+   implementation do not return domain
+ - i18n for `Select key media`     
 
 ## [5.5.16] - 2020-07-19
 ### Added