+++ packages/ub-pub/CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+
+## [5.5.17] - 2020-07-26
+### Fixed
  - fixed missed localization for `Select device` caption on `@ub-d/iit-sign-web/IitReadPk.vue` form
  - prevent sets of `connection.domain` to null in case `onGotApplicationDomain` handler is passed to `AsyncConection.connect` and
    implementation do not return domain